Remove 2D tests and split 1D and 3D tests
