+++ tests/CMakeLists.txt
@@ -1,34 +1,29 @@
 set(_tests_in_build_tree ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/..)
 
-configure_file(
-  ${CMAKE_CURRENT_SOURCE_DIR}/test_gauss.py
-  ${_tests_in_build_tree}/test_gauss.py
-  COPYONLY
-  )
-configure_file(
-  ${CMAKE_CURRENT_SOURCE_DIR}/test_mra.py
-  ${_tests_in_build_tree}/test_mra.py
-  COPYONLY
+list(APPEND _tests
+  basis
+  mra_1d
+  mra_3d
+  gauss_1d
+  gauss_3d
   )
 
-# Define tests
-add_test(
-  NAME
-    mra
-  COMMAND
-    COMMAND ${PYTHON_EXECUTABLE} -m pytest -x -rws -vv ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/../test_mra.py
-  )
-add_test(
-  NAME
-    gauss
-  COMMAND
-    COMMAND ${PYTHON_EXECUTABLE} -m pytest -x -rws -vv ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/../test_gauss.py
-  )
-
-install(
-  FILES
-    test_mra.py
-    test_gauss.py
-  DESTINATION
-    ${PYMOD_INSTALL_FULLDIR}/..
-  )
+foreach(_test IN LISTS _tests)
+  configure_file(
+    ${CMAKE_CURRENT_SOURCE_DIR}/test_${_test}.py
+    ${_tests_in_build_tree}/test_${_test}.py
+    COPYONLY
+    )
+  add_test(
+    NAME
+      ${_test}
+    COMMAND
+      COMMAND ${PYTHON_EXECUTABLE} -m pytest -x -rws -vv ${PROJECT_BINARY_DIR}/${PYMOD_INSTALL_FULLDIR}/../test_${_test}.py
+    )
+  install(
+    FILES
+      test_${_test}.py
+    DESTINATION
+      ${PYMOD_INSTALL_FULLDIR}/..
+    )
+endforeach()
